requirements.txt
environs~=9.5.0
python-telegram-bot~=13.11
google-cloud-dialogflow~=2.11.0